Move PluginMonitor out of TLuaEngine
