+++ CMakeLists.txt
@@ -120,6 +120,7 @@ set(BeamMP_Sources
     include/TScopedTimer.h src/TScopedTimer.cpp
     include/SignalHandling.h src/SignalHandling.cpp
     include/ArgsParser.h src/ArgsParser.cpp
+    include/TPluginMonitor.h src/TPluginMonitor.cpp
     include/Environment.h
 )
