+++ src/TLuaEngine.cpp
@@ -15,8 +15,7 @@
 
 TLuaEngine* LuaAPI::MP::Engine;
 
-TLuaEngine::TLuaEngine()
-    : mPluginMonitor(fs::path(Application::Settings.Resource) / "Server", *this, mShutdown) {
+TLuaEngine::TLuaEngine() {
     Application::SetSubsystemStatus("LuaEngine", Application::Status::Starting);
     LuaAPI::MP::Engine = this;
     if (!fs::exists(Application::Settings.Resource)) {
@@ -35,7 +34,7 @@ TLuaEngine::TLuaEngine()
         }
         Application::SetSubsystemStatus("LuaEngine", Application::Status::Shutdown);
     });
-    Start();
+    IThreaded::Start();
 }
 
 TEST_CASE("TLuaEngine ctor & dtor") {
@@ -987,70 +986,4 @@ bool TLuaEngine::TimedEvent::Expired() {
 
 void TLuaEngine::TimedEvent::Reset() {
     LastCompletion = std::chrono::high_resolution_clock::now();
-}
-
-TPluginMonitor::TPluginMonitor(const fs::path& Path, TLuaEngine& Engine, std::atomic_bool& Shutdown)
-    : mEngine(Engine)
-    , mPath(Path)
-    , mShutdown(Shutdown) {
-    if (!fs::exists(mPath)) {
-        fs::create_directories(mPath);
-    }
-    for (const auto& Entry : fs::recursive_directory_iterator(mPath)) {
-        // TODO: trigger an event when a subfolder file changes
-        if (Entry.is_regular_file()) {
-            mFileTimes[Entry.path().string()] = fs::last_write_time(Entry.path());
-        }
-    }
-    Start();
-}
-
-void TPluginMonitor::operator()() {
-    RegisterThread("PluginMonitor");
-    beammp_info("PluginMonitor started");
-    while (!mShutdown) {
-        std::this_thread::sleep_for(std::chrono::seconds(3));
-        std::vector<std::string> ToRemove;
-        for (const auto& Pair : mFileTimes) {
-            try {
-                auto CurrentTime = fs::last_write_time(Pair.first);
-                if (CurrentTime != Pair.second) {
-                    mFileTimes[Pair.first] = CurrentTime;
-                    // grandparent of the path should be Resources/Server
-                    if (fs::equivalent(fs::path(Pair.first).parent_path().parent_path(), mPath)) {
-                        beammp_info("File \"" + Pair.first + "\" changed, reloading");
-                        // is in root folder, so reload
-                        std::ifstream FileStream(Pair.first, std::ios::in | std::ios::binary);
-                        auto Size = std::filesystem::file_size(Pair.first);
-                        auto Contents = std::make_shared<std::string>();
-                        Contents->resize(Size);
-                        FileStream.read(Contents->data(), Contents->size());
-                        TLuaChunk Chunk(Contents, Pair.first, fs::path(Pair.first).parent_path().string());
-                        auto StateID = mEngine.GetStateIDForPlugin(fs::path(Pair.first).parent_path());
-                        auto Res = mEngine.EnqueueScript(StateID, Chunk);
-                        // TODO: call onInit
-                        mEngine.AddResultToCheck(Res);
-                    } else {
-                        // TODO: trigger onFileChanged event
-                        beammp_trace("Change detected in file \"" + Pair.first + "\", event trigger not implemented yet");
-                        /*
-                        // is in subfolder, dont reload, just trigger an event
-                        auto Results = mEngine.TriggerEvent("onFileChanged", "", Pair.first);
-                        mEngine.WaitForAll(Results);
-                        for (const auto& Result : Results)  {
-                            if (Result->Error) {
-                                beammp_lua_error(Result->ErrorMessage);
-                            }
-                        }*/
-                    }
-                }
-            } catch (const std::exception& e) {
-                ToRemove.push_back(Pair.first);
-            }
-        }
-        for (const auto& File : ToRemove) {
-            mFileTimes.erase(File);
-            beammp_warn("file '" + File + "' couldn't be accessed, so it was removed from plugin hot reload monitor (probably got deleted)");
-        }
-    }
 }
